Add asynchronous functionality
Middleware manager enhanced to allow layers to be run asynchronously in the order they were add to the manager. Manager’s API has been updated to include `applyAsync` and `applyForKeyAsync`. Both methods have a similar signature as `apply` and `applyForKey` expect that they return a promise and an additional `reject` value is supplied after `next`.